+++ src/ProgressOnderwijsUtils/Utils.cs
@@ -11,17 +11,6 @@ using ProgressOnderwijsUtils.Collections;
 
 namespace ProgressOnderwijsUtils
 {
-    public static class DisposableExtensions
-    {
-        public static T Using<TDisposable, T>(this TDisposable disposable, Func<TDisposable, T> func)
-            where TDisposable : IDisposable
-        {
-            using (disposable) {
-                return func(disposable);
-            }
-        }
-    }
-
     public static class Utils
     {
         /// <summary>
